neural_agents/requirements.txt
# Core Dependencies
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.17
pydantic==2.5.2
openai==1.3.7

# Web Framework
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0

# Data Processing
numpy==1.24.3
pandas==2.0.2

# Visualization
matplotlib==3.7.1
plotly==5.18.0
networkx==3.1
graphviz==0.20.1
ipywidgets==8.0.6

# Testing and Development
pytest==7.4.0
black==23.7.0
isort==5.12.0

# Embeddings and Vector Stores
chromadb==0.4.18
sentence-transformers==2.2.2 